cusco: overlay: Enable smooth display with 120Hz
There's no point to keep using custom pref to set minimum refresh rate
when AOSP smooth display feature behaves in the exact same way in
our case.

Set default peak refresh rate to 60, as this is the same value when the user
turns on and off the smooth display setting.
If its turned on, it will go to 144 Hz.

Change-Id: I497918a9c9687443a9438fb6291769f6e51d9379